+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/config/Config.java
@@ -1,8 +1,8 @@
 package com.github.khanshoaib3.minecraft_access.config;
 
 import com.github.khanshoaib3.minecraft_access.MainClass;
-import com.github.khanshoaib3.minecraft_access.config.feature_config_maps.FallDetectorConfigMap;
-import com.github.khanshoaib3.minecraft_access.config.feature_config_maps.OtherConfigsMap;
+import com.github.khanshoaib3.minecraft_access.config.config_maps.FallDetectorConfigMap;
+import com.github.khanshoaib3.minecraft_access.config.config_maps.OtherConfigsMap;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import net.minecraft.client.MinecraftClient;
