+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/config/config_menus/CameraControlsConfigMenu.java
@@ -1,6 +1,8 @@
-package com.github.khanshoaib3.minecraft_access.config;
+package com.github.khanshoaib3.minecraft_access.config.config_menus;
 
 import com.github.khanshoaib3.minecraft_access.MainClass;
+import com.github.khanshoaib3.minecraft_access.config.ConfigMap;
+import com.github.khanshoaib3.minecraft_access.utils.BaseScreen;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.resource.language.I18n;
 import net.minecraft.text.Text;
