+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/features/FluidDetector.java
@@ -1,7 +1,7 @@
 package com.github.khanshoaib3.minecraft_access.features;
 
 import com.github.khanshoaib3.minecraft_access.MainClass;
-import com.github.khanshoaib3.minecraft_access.config.feature_config_maps.FluidDetectorConfigMap;
+import com.github.khanshoaib3.minecraft_access.config.config_maps.FluidDetectorConfigMap;
 import com.github.khanshoaib3.minecraft_access.utils.PositionUtils;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
